Removed dependency cycle between the analyzer, engine, and filter packages.
Removed dependency cycle between the analyzer, engine, and filter packages by moving the analyzer package under the engine package directly instead of having it live in the sta package since the ContextAnalyzer class is used in the Processor class and refers to the VertexFilterFactory class.  The ContextAnalyzer class could just as easily live directly in the engine package and could be moved there if desired.